requirements.txt
Pillow
numpy
tensorflow>=2.0.0
tensorflow-hub
streamlit